creating table -anneee_ecole -annee_ecole_classe -annee_ecole_classe_enseignant - Suppression des fichiers de type ApiService

+++ src/main/java/sen/school/primaire/annee/AnneeApiService.java
@@ -13,7 +13,7 @@ public class AnneeApiService {
     @Autowired
     AnneeRepositoryService anneeRepositoryService;
 
-    public Optional<Annee> findById(int id) {
+    public Optional<Annee> findById(Long id) {
         return anneeRepositoryService.findById(id);
     }
