+++ src/main/java/sen/school/primaire/ecole/EcoleRepository.java
@@ -3,7 +3,7 @@ package sen.school.primaire.ecole;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sen.school.primaire.annee.Annee;
 
-public interface EcoleRepository extends JpaRepository<Ecole, Integer> {
+public interface EcoleRepository extends JpaRepository<Ecole, Long> {
 
 
 }
