+++ src/main/java/sen/school/primaire/eleve/EleveRepository.java
@@ -2,7 +2,7 @@ package sen.school.primaire.eleve;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EleveRepository extends JpaRepository<Eleve, Integer> {
+public interface EleveRepository extends JpaRepository<Eleve, Long> {
 
 
 }
